requirements.txt
python-dotenv
aiogram==2.5
openai==0.28.1
streamlit==1.29.0
google.generativeai